refactor(utils): move getAPIPathRoot, sendGetRequest out to a common utils file. Adds sendPutRequest to utils.
